regulator: add Rockchip rk808 support
This driver provides control of following power management ICs by
Rockchip: RK805 RK809, RK817, RK808 and RK818 as well as a power-off
handler in the MFD driver.

The other MFD cells are exported as well, but no drivers are added for
them here. The code is based on an older version of the Linux driver,
with less abstractions that are not yet synced to barebox.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>